+++ src/odig_odoc.mli
@@ -13,7 +13,8 @@ open Odig_support
 
 val gen :
   Conf.t -> force:bool -> index_title:string option ->
-  index_intro:B00_std.Fpath.t option -> pkg_deps:bool -> tag_index:bool ->
+  index_intro:B00_std.Fpath.t option -> index_toc:B00_std.Fpath.t option ->
+  pkg_deps:bool -> tag_index:bool ->
   Pkg.t list -> (unit, string) result
 (** [gen c ~force ~index_intro ~pkg_deps ~tag_index pkgs]
     generates API reference for packages [pkgs].
@@ -22,6 +23,8 @@ val gen :
        with the list of packages.}
     {- [index_intro] if specified is an mld file to
        define the introduction of the page with the list of packages.}
+    {- [index_toc] if specified is an mld file to
+       define the table of contents for the page with the list of packages.}
     {- [pkg_deps] if [true] dependencies of [pkgs] are also generated.
        If [false] only [pkgs] are generated which may lead to broken
        links in the output.}